+++ packages/rpc-base/CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.1.3 (2018-08-16)
+
+Fix CommonJS export.
+
 ## v0.1.2 (2018-07-09)
 
 Export as CommonJS and ESM.