+++ packages/rpc-handler/CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.2.0 (2018-08-17)
+
+Add `onHandlerError` option.
+
 ## v0.1.0 (2018-07-31)
 
 First release.